removed parseInt with parseFloat

+++ index.js
@@ -81,8 +81,8 @@ async function run() {
     app.post('/toyCars', async (req, res) => {
       const toyInfo = req.body
       // console.log(toyInfo)
-      toyInfo.price = parseInt(toyInfo.price)
-      toyInfo.rating = parseInt(toyInfo.rating)
+      toyInfo.price = parseFloat(toyInfo.price)
+      toyInfo.rating = parseFloat(toyInfo.rating)
       const result = await toyCarsCollection.insertOne(toyInfo)
       res.send(result)
     })
@@ -98,8 +98,8 @@ async function run() {
         $set: {
           picture: toyInfo.picture,
           name: toyInfo.name,
-          price: parseInt(toyInfo.price),
-          rating: parseInt(toyInfo.rating),
+          price: parseFloat(toyInfo.price),
+          rating: parseFloat(toyInfo.rating),
           sellerName: toyInfo.sellerName,
           sellerEmail: toyInfo.sellerEmail,
           description: toyInfo.description,
